dfilter: Extra value_string checking and comments
It's possible to register multiple fields to the same abbreviation.
We don't actually enforce field compatibility, but even if we did,
FT_FRAMENUM is considered compatible with other integer fields.
So it's theoretically possible to register a integer field with a
value string to the same abbreviation as a FT_FRAMENUM with a non-NULL
strings - FT_FRAMENUM overloads that field to store ft_framenum_type,
and treating it as a value_string *will* crash. (I don't believe
we have any such fields currently.)

Add a couple other comments regarding treatment of value strings -
we shouldn't actually optimize tests against strings that represent
value strings in cases where multiple fields are registered to the
same abbreviation and don't share exactly the same value string.
(See #19111 for a specific example.)