cllog: Avoid a stack read overrun.
If parseFieldLength received a length of greater than 8 and parseFieldData was
not subsequently called, the memcpy at the end of cllog_read_common could read
beyond the end of the eight-byte buffer in logEntry.data on the stack.  This
condition could arise if the column header in a malicious input file stipulated
that the Length field appeared _after_ the Data field, or perhaps if the Data
field was absent.

The result would most likely be just some garbage data from the stack appearing
in the packet contents, but there's a chance it might also cause a crash if
length is large enough to read beyond the top of stack.

Clean up a related magic number "8" and an unnecessary loop counter.

Ping #19937