ftypes: Unify some of the "compatible types" handling
The ftypes are divided into equivalence classes of compatible types.
Multiple fields can be registered to the same abbrevitation so long
as all the fields have the same type. The handling is inconsistent
and scattered, however.

README.dissector gives a list of the types and one set of classes.
proto.c has an implementation, used only when hf conflict check is
turned on when compiling.
ftypes.c has an implementation, used in Lua for registering ProtoFields.
The most consequential version is in the DFilter semantic check, which
determines whether a filter is considered valid at all based on which
fields can be compared. It also determines the type to use when
converting various literal tokens to fvalues. (Note that this does
not always work properly, both for arithmetic operation with multiple
operands, and for fields that multiple compatible types registered,
but where one type is more restrictive in what literals it accepts.)
None of these are quite the same.

Have proto.c's hf conflict check use the same function in ftypes.c
used by Lua registration. Add some types to that function so that
it splits the difference between README.dissector and the dfilter
semcheck version. (The differences are with the time types and
whether the OID group is part of the BYTES group.) Add some
comments to the implementations where they differ. Move FT_AX25
in the README to the STRING group, as already done in the dfilter
semcheck as part of commit 92d958ad45a500d8074c7ca9f393eb41619a5df7.