FAQ: Minor cleanups; Remove Yahoo Messenger content.
The Yahoo Messenger service has been defunct since August 2016
(according to Wikipedia), so it seems unnecessary to describe a Yahoo
Messenger protocol oddity in the FAQs.

Genus names are conventionally shown with an initial capital letter; I
guess this rule extends to mascot shark "species". :-)

Fix an incorrect reference which resulted from the movement of the
Ethereal naming discussion to the end of the document.

Remove a stray backtick.