asn2wrs: Don't use implicit -1 choice tags in value strings
When parsing a CHOICE type with untagged choices that are themselves
CHOICEs (and implicit tagging instead of AUTOMATIC tagging), we
set the tag as a special value of -1 so that dissect_ber_choice()
will know to try various choices at the next level.

(This is an internal tag value not something encoded in the packet.
The tag as actually encoded is from the type ultimately chosen,
and should be unique once all the choice levels are traversed,
see ITU-T X.680 29.3 and the examples in 29.9.)

We add to the tree an item that represents which choice was taken;
rather than adding a single item that reflects the ultimate choice
we add items for each level of choice. Those items get value strings -
the value string values can either be directly from the tags used in
the CHOICE, or can be assigned incremently from 0 enumerating the
choices. We shouldn't use these special tag numbers of -1 in a value
string for the choice; we should always fall back to assigning values
for each choice enumerated from 0.

Most of the time our existing checks catch this either at least one
of the tags will be something other than a CHOICE, and thus will have a
different tag class, or AUTOMATIC tagging is used.) However, if all the
choices of a CHOICE are themselves a CHOICE and IMPLICIT tagging is used
the current checks don't catch it.

IPAddress from GPRSCDR is such a type. This change causes
iPBinaryAddress to have a value of 0 and iPTextRepresentedAddress to
have a value of 1, instead of both having a value of -1 (4294967295),
eliminating a large number of warnings of the form:

 [Epan WARNING] epan/proto.c:9249 -- tmp_fld_check_assert(): Field 'iPAddress' (gprscdr.iPAddress) has a conflicting entry in its value_string: 4294967295 is at indices 0 (iPBinaryAddress) and 1 (iPTextRepresentedAddress)

when ENABLE_CHECK_FILTER is defined (as in the Clang + Code Checks job)

This allows unique filtering on gpsrcdr.iPAddress and other fields using
that ChoiceType, like gprscdr.ggsnAddress.