dfilter: FT_GUID is not a bytes type
FT_GUID is not a bytes type, and is not compatible with the FT_BYTES
family. This has been the case since 923855bb20845f55472dfe8e62b7b43660f2f63b
in order to support both big and little endian GUIDs.

This prevents:
[GLib CRITICAL] -- g_bytes_compare: assertion 'bytes2 != NULL' failed
(and always returning "equals") when comparing a member of the bytes
equivalence class on the left hand side to a FT_GUID on the right hand
side of a filter comparison (e.g. eth.addr == smb2.client_guid), and
when the FT_GUID is on the left side, doing a memcmp with type-punning
of the byte type to a e_guid_t.

Also, use guid_cmp from "epan/guid-utils.h" to compare the e_guid_t
structs element by element instead of using memcmp.