text2pcap: Regex parser should not require a newline.
Using g_regex_match_full allows us to avoid the need to NUL-terminate
the input file, which is nice because that NUL is being written beyond
the defined length of the mmap'd region -- it could cause a crash or
undefined behavior if the input file is an exact integer multiple of
the system's page size and the NUL is written to the following page.

There's also no need to insist upon a newline at the end of the file.
All of the parsing code is already written to not depend on the fields
being delimited in any specific way, so this test appears superfluous.

Clean up a _potential_ infinite loop if we ever reached the path where
the "data" field was not found.  Failing to call g_match_info_next
and continuing from the top of the loop would prevent it from ever
advancing from that point.  I'm not convinced that it is truly
possible to reach that path with G_REGEX_MATCH_NOTEMPTY specified, but
it is easy enough to make it safe.

Finally, add very basic unit tests for the regex functionality of
text2pcap, including a case which lacks a newline at the end of the
file.  The test is based upon the minimal example in the text2pcap
utility's "--help" text.

Fixes #20245.