More help output cleanup, making the output more alike.
Note that the -B argument is in units of MiB.

Put --update-interval in a "Capture display" section for TShark, as it
affects how the output is displayed (printed).  Don't duplicate it for
Wireshark, and put it at the end of the list.

--only-protocols takes a list of protocol, not a single protocol; make
the Wireshark output match the TShark output.


(cherry picked from commit 9a24ae406e1d5fd0a369c6e70d776075687af0c9)

Co-authored-by: Guy Harris <gharris@sonic.net>