ftypes: All integer types are compatible
On the dfilter side, all the integer types are compatible since
commit 4c975b770e6bc6d16909c76954877d54d8f1f47b. They don't need
to be restricted from being registered with the same abbreviation.
We already have a number of fields that are registered this way,
and they work as well as other integer fields of different sizes
registered to the same abbreviation.

There are some existing issues with the semantic check refusing to
convert literals for comparison when integer types of different sizes
have the same abbreviation and the smaller field is registered last,
related to #19618 and #18960. We only test the range based on the last
field registered. However, that problem occurs with integer types that
are are allowed to be registered to the same abbreviation, like FT_UINT16
and FT_UINT32 and needs to be fixed anyway.

Change one Lua test since FT_INT16 and FT_UINT16 are no longer
incompatible.