DHCP: Don't try to finalize a composite TVB with all zero length TVBs
Zero length TVBs aren't added to composite TVBs. Trying to finalize
a composite TVB with no members throws a dissector error. (The code
can't really handle it in either case and might crash.)

So don't create the composite TVB until we get a nonzero option length.
If all the options have zero length, there won't be a composite TVB,
and we can just mark that with an expert info.

Fix #20260