ftypes: Allow comparing unsigned and signed integers
Allow signed and unsigned integer ftypes to be compared. We
always currently call the compare function of the left hand side
type. If the right hand side fails to convert to the appropriate
type due to integer underflow, we know the LHS is greater. If
the RHS fails to convert due to overflow, we know the LHS is lesser.

Ping #18960

(There are some other corner cases in that issue, plus we might want
to warn/fail on comparisons to constants that are outside type limits
and always true or always false.)