Provide a macro for code specific to fuzz testing.
There's a quasi-standard in the industry for supporting fuzz testing
through use of the macro "FUZZING_BUILD_MODE_UNSAFE_FOR_PRODUCTION" to
conditionalize code for maximum fuzzing efficiency -- see
https://llvm.org/docs/LibFuzzer.html#fuzzer-friendly-build-mode .

Common uses for this macro would be to replace a nondeterministic
process (such as fetching a random number or determining the
time-of-day of execution) with a substitute deterministic process, or
to bypass the validation of checksums or digital signatures where they
would otherwise hamper fuzz-test coverage.

As the macro name suggests, it is never defined for regular
"non-fuzzing" builds, so whichever alterations are made under the
control of this macro should have no effect on regular usage.

This change simply adds the macro and wires it up in CMake so that it
is defined in config.h iff either of the ENABLE_FUZZER or OSS_FUZZ
CMake flags are set.