addr_resolv: Load ethers file when initializing name lookups
Instead of parsing the entire ethers file each time a new EUI-48 address
is encountered, read the file when initializing name lookups and use the
cached results. Reading the text file is always linear in the length of the
file, so this is O(N) instead of O(N*M).

The original behavior allowed UN*X like updating of the lookup if the
/etc/ethers file was changed on the fly, but that certainly wasn't worth
the performance hit (#19915). Caching the results and updating whenever
opening a new capture file or changing configuration profiles should be
sufficient.

Update the documentation in the WSUG to reflect when the various name
resolution files are read, as the information is a bit inaccurate or
out of date.

Note that the ipxnets and vlans files also do on-demand lookup. IPX network
addresses are rare these days, and vlan name lookup is disabled by default,
but we still might want to change those to preloading the files like the
others too.