wustil: Add David Gay's dtoa for shortest sufficient double representation
For filtering with equality on FT_DOUBLEs, we need to produce a string
value which, when converted back to a double, will compare as equal to
the original value.

That can be achieved by using a printf %g conversion with DBL_DECIMAL_DIG
(as g_ascii_dtostr does), and that's what the FTREPR_DFILTER string
representation currently does. Note that DBL_DECIMAL_DIG is not the
same as DBL_DIG, which is the number of significant digits guaranteed
not to change when converting a string to a double and back to a
string.

Using only DBL_DIG significant digits is not sufficient; for example, the
IEEE 754 double precision binary floating point number closest to 1/7
(0x1.2492492492492p-3 in hexadecimal floating point notation) converted
with DBL_DIG siginficant digits yields the string "0.142857142857143".
That string if converted back to a double produces 0x1.2492492492497p-3,
which does *NOT* compare as equal. A longer string, such as
"0.14285714285714285", is necessary for round trip conversion back to
a double to work.

However, there are many strings with between DBL_DIG and DBL_DECIMAL_DIG
significant digits that convert to the same double; the one with the
largest number of trailing zeros (that thus becomes shortest after
removing the trailing zeroes) may not be the numerically closest to
the exact floating point number and thus not the one derived via the
direct algorithm.

E.g., the doubles precisely expressed as 0x1.999999999999ap-3 and
0x1.3333333333333p-2 in hexadecimal floating point notation are
closer to 0.20000000000000001 than 0.2 and 0.29999999999999999 than to
0.3, respectively, even though both decimals convert to the same
IEEE 754 binary double.

David M. Gay's dtoa implementation (https://www.netlib.org/fp/) is a
long standing implementation used in a variety of projects (such as Python -
https://github.com/python/cpython/blob/main/Python/dtoa.c ).

Remove the strtod implementation, as we don't need it (it was added
in C99 and C++11). Change a few things to deal with compiler warnings
(such as the comments for FALLTHROUGH in case statements and casts.)
There's probably some more ifdef'ed code that could be removed as well,
as Python did.

There are a number of C++ libraries that implement apparently faster
algorithms, but few other C implementations. Also note that C++17
added std::to_chars that produces the same result on floating point
numbers (though the difficulty in implementation meant that this
was one of the last C++17 features added to most C++ standard library
implementations.)

Add its MIT-like license (see https://spdx.org/licenses/dtoa.html)
to licensecheck.

Ping #16483

This achieves the goal of presenting the same shortest result in
the protocol tree for FT_DOUBLEs in the tree as well as in the
filter and serializing.

Note this only produces the shortest result for *doubles*; extending
this to IEEE 754 single precision floating point requires using the
generalized version of the library (which may be necessary.)