dfilter: value string matching with multiple hfinfo same abbrev
It's possible to register fields to the same abbreviation. They
*should* all be compatible, but that's not enforced. Even if
compatible, different fields can have different value strings.
(vlan.priority is an example.)

We can optimize the filter to a numerical comparison if and only
if every hfinfo for the given abbreviation has the same value (and
only that value) mapped to the string. Otherwise, the semantic
check can allow a string comparison if at least one of the hfinfos
for a given abbreviation has a value string that has that string
present as a value. That also means not failing immediately if
one hfinfo can't convert, until we check others.

See #19111.

Note this doesn't affect the Filter Expressions Dialog, which probably
only suggests possibilities from one value string instead of all of
them.