tls: Workaround for missing handshake secret, app secret present
Mozilla NSS appears to have a bug when the ML-KEM key exchange support
was added so that the HANDSHAKE_TRAFFIC_SECRETs are not being written
to the SSLKEYLOGFILE even though the other secrets like the application
TRAFFIC_SECRET_0 are. Since the Finished message is always encrypted
with the handshake key, we'll never try to switch to the app traffic
key, even if we have it and could decrypt the TLS records that are
encrypted using it.

To work around such situations, in TLS 1.3 if we fail to retrieve
the handshake traffic key, proceed to go ahead and try to retrieve
the appplication traffic key and store it for use against future TLS
records. We won't decrypt records up to the Finished, but payload later
will be decrypted.

We don't need to do this for QUIC, which already retrieves the
appropriate key based upon the type indicated in the packet header
(Handshake vs Protected Payload.)

Workaround #20240