in_cksum: Prefer returning 0x0000 to 0xFFFF from in_cksum_shouldbe
The one's complement math in in_cksum_shouldbe can't distinguish
between the two zero representations +0 (0x0000) and -0 (0xFFFF).
The current algorithm is one that favors producing 0xFFFF, but
for nearly all protocols and situations 0x0000 is what the checksum
"should be."

0xFFFF is only the preferred result when all other bits over which
the checksum is calculated (except the checksum itself) are zero.
Unfortunately in_cksum_shouldbe doesn't know that; fortunately, all
zero is not a valid packet for most protocols that use the checksum,
whether because of a version field, length field, or something else
in the packet (or pseudoheader.) So prefer 0x0000, which is generally
the correct answer.

One of the few exceptions is ICMP, where a legal Echo Reply with ID
0, sequence number 0, and no or all zero data is legal. Perform
the ICMP checksum processing in a different way that allows us to
distinguish the cases. Note this also will produce an expert info
for cases where 0xFFFF is substituted for 0x0000 - ICMP probably
isn't subject to incremental updates as mentioned by RFC 1624, but
we probably want to flag those cases in some way anyway. We will
still match those in request-reply tracking because negative zero
is mapped to positive zero in the match key.

Note there are networking stacks that insert the opposite zero in both
cases. (Linux currently does for all zero ICMP Echo Requests.) Our
normal method of performing Internet checksums always accepts 0xFFFF
where 0x0000 is correct (but not vice versa) due to verifying by including
the checksum bits in the check. For TCP we have a special expert warning
for this referencing RFC 1624; we could do this in more generality.

Fix #16334