TCP: Multisegment PDUs don't contain keep-alives
Per RFC 9293  3.8.4. TCP Keep-Alives,

"Keep-alive packets MUST only be sent when no sent data is outstanding"
and "there is no new or unacknowledged data to be sent." Keep-alives
are distinguished from retransmissions and zero window probes.
They SHOULD be sent with no data, but for compatibility some network
stacks are configured to send a garbage octet.

Don't add the garbage octet, if present, to a multi-segment PDU for some
of the same reasons that we don't add retransmissions.

In fact, we can go further, as the above means that a multi-segment PDU
cannot span a keep-alive boundary. (Unlike retransmissions and zero window
probes.) If we have a keep-alive, and also an unfinished MSP that started
at the same sequence number as this keep-alive, then the previous MSP is
in fact also a keep-alive garbage octet that was sent to the next
dissector (presumably configured for the port) and incorrectly
determined to be the start of a MSP. This happens if the first packet
for a TCP stream in a capture is a keep-alive segment with a garbage
octet, because we cannot determine that the first segment is a
keep-alive with no prior sequence number.

Fix #20287