Kerberos: add AD-CAMMAC and AD-AUTHENTICATION-INDICATOR
RFC-8129 defines an "authentication indicator" ad-type (97), which in
turn uses the CAMMAC container ad-type (96) defined in RFC-7751. Add
these to the Kerberos dissector.

It would be nice to also verify the checksums in a CAMMAC container
using keys from the user-supplied keytab, but that will require more
custom code, similar to the existing code which verifies signatures in
win2k-pac elements. Perhaps later.