tetra: Remove unusued hf with duplicate abbreviation
hf_tetra is a FT_NONE that is manually defined and not actually
used or added to the tree anywhere. It has the same abbreviation
as a FT_UINT32 field that is produced by the ASN.1 and can actually
be added to the tree. Remove the unused field to eliminate a warning
about a field existing multiple times with incompatible types.