GNSS: Use different abbrevs for expert infos than the invalid fields
Add a suffix to the expert infos' filter abbreviations so that they
can be filtered for separately from the integer fields on which they're
based. This fixes some warnings about incompatible types being
registered to the same filter abbreviation.

Each of these are only used once. There's no point in using
expert_add_info_format() with a const string slightly different from
the static string used when registering the expert infos.
Just register them with the constant string that will be used.