WASSP: Make TLV type filterable again
The sub TLV type being added later as a FT_NONE with the same
filter abbreviation made it impossible to filter by TLV type.
Use a unique filter abbreviation for the field. Also, add the
sub type as a FT_UINT16 with a custom format (it's selected from
one of several tables, and possibly should be multiple hf infos)
instead of as a FT_NONE so that the sub types can also be filterable.