BER: Allow unknown OCTETSTRINGs to be filtered
BER registers two fields for unknown OCTETSTRINGS, one for when the
unknown OCTET STRING looks like it could be BER encoded data, and
a preference is enabled to try to decode it that way. They both use
the same filter abbreviation.

For compatibility, instead of the latter one being a FT_NONE, make
it a FT_BYTES with BASE_NO_DISPLAY_VALUE. That makes everything look
the same as before, except that the types are compatible so that they
can be filtered (and changing the preference doesn't change the
filter result.)

This removes another case of fields with the same filter abbreviation
being non-compatible.