IEEE1722: Don't reuse can.reserved in ACF CAN
ACF CAN reserves 3 high bits from a 32 bit word that stores a
29 bit CAN identifier. Don't name these bits (a FT_UINT32) with
the "can.reserved" abbreviation, because that's already used by
SocketCAN for a FT_BYTES, which is not a compatible field type.
(SocketCAN uses those three bits for the eff, rtr, and esi bits,
which are placed in a different location in IEEE 1722 ACF CAN.
SocketCAN has reserved bytes elsewhere.) Just use a unique name.

There is some logic in having fields that correspond to the same
thing have common abbreviations, but that doesn't apply to the
reserved bits and bytes which are different in the various encodings.