text2pcap: Handle preamble tokens that look like offsets or bytes
If we encounter a byte token where we don't expect a byte, treat
it as text and add it to the preamble. It might be a day or month.
If we encounter an offset token that doesn't match, treat it as
text and add it to the preamble. It could be a year or similar.

There's still some ambiguity, which could be reduced by only allowing
the mailfwd character ('>') and blanks to preceed the offset on a line,
requiring preamble information to appear on separate lines.

The parse_num function will reject too large offsets, but large
numerical strings can appear in preambles. Just limit the size of
the offset token to 8 hex characters. (Even in octal, 8 characters
gives much larger offsets than allowed in a single packet by
text_import: WTAP_MAX_PACKET_SIZE_STANDARD is 262144.)

Also handle multiple mailfwd characters before an offset (with no blanks
in between), not just one.

Fix #17140. Ping #16193