text_import: Improve the extra byte check
When seeing an unexpected offset, we should only truncate if
the offset is greater than the last line start as well as less
than the expected value. We know the last line start offset
(it's used in the similar identify_ascii processing), and truncating
is only correct in that narrow window. (We only add bytes if a line
starts with the correct offset, so extra bytes added from the text
dump have to be within the current line.)

This check largely overlaps with the -a / "Identify ASCII" flag,
though it would have some extra utility if we ever support
multiple byte groups as some hex dump programs produce.

Clean up a few comments, add a missing statement to set packet_start
to 0 inside INIT on bad offsets. Don't both adding to the preamble
in no offset mode after we've started a packet.

This prevents most false detections of preamble strings as offsets.
(#17140)