netscaler: slight code cleanup.
There's a section of nstrace_dump() that tests the file version and,
regardless of the version, calls the same two routines in the same
fashion.  Those routines also check the file version and do different
things; if the version is something unexpected, they fail, so the check
for a sane version in nstrace_dump() is unnecessary.

Just call those routines without checking the version.