columns: Add "Z" suffix to UTC times
ISO 8601 specifies that times given without time zone information are
assumed to be in local time. Custom columns based on FT_ABSOLUTE_TIME
already add time zone information, but the frame time columns such
as "Time (format as specified)", and "UTC time" do not.

Add a Z suffix for UTC, per ISO 8601. Keep the local time formats
without a time zone suffix, as that is permissible. This makes the
time column output less dependent on out of band information and able
to be used in other situations (such as inputting to text2pcap - see #17132.)