wiretap: have wtap_init_rec() clean the wtap_rec's buffer.
That's part of the process of resetting a wtap_rec for reuse with a new
record; we were already doing that right after calling wtap_init_rec(),
but this means that wtap_init_rec() does all the work.

Note in a comment that it might have to reset other buffers.

In the comment before wtap_init_rec(), more fully and accurately explain
what it does.