Buffer: just call ws_buffer_clean() in ws_buffer_remove_start().
We now have a routine whose sole job is to reset a Buffer; call that if
ws_buffer_remove_start() has removed everything in the buffer.

Add a comment indicating that's what we're doing.

While we're at it, wrap one #ifndef SOME_FUNCTIONS_ARE_DEFINES/#endif
pair around the routines that may be implemented as macros in buffer.h,
rather than wrapping each one individually.  That matches what we do in
the header.