Drop support for pre-1.0 libpcap on non-Windows
Require pcap_create(). That first appeared in libpcap 1.0, which
was released 2008-10-17. That also ensures that the entire libpcap
1.0 and earlier API is present, so we can remove some checks and
ifdefs.

On Windows the compile time checks have always been skipped and
the functions checked for at run time, so this really just removes
all the ifdefs but the run time checks are still there. A subsequent
commit could remove the run time checks and just refuse to load the
dll if pcap_create and other 1.0 APIs are not present.

Note WinPcap 4.1.3 fails to export three functions (pcap_statustostr(),
pcap_can_set_rfmon(), and pcap_set_rfmon()), so even if we do require
the rest of the libpcap 1.0 APIs on Windows, we might still want to mark
those as optional in Windows and work around their absence.

Ping #20261