hex dumps: Add timestamp option
It can be useful when producing hex dumps to be used as input to text2pcap
to include the frame timestamp, while minimizing extraneous information
in the preamble. On tshark, this can be achieved via recipes like

    -P -o gui.column.format:"Time","%Yt"

(print packet summary regardless of other options, set the columns
to a single Absolute Date and Time column.)

On Wireshark this could be achieved similarly by temporarily hiding
columns other than a time column (and possibly changing the Time
Display Format.)

Add an option to print the time to the --hexdump long options added
in b5f89dbe2db357a0f4a08de7d00fdeed7ca2780c and add it to the GUI.
This should be somewhat easier to use.

Fix #17132