Update our Stratoshark artwork
Update the following icons:

- The Windows application icon (stratoshark.ico)
- The macOS application and volume icons (Stratoshark.icns and
  ssiconvol.icns)
- The Qt icon resources (ssicon[1-9]*.png and ssiconcap[1-9]*.png)
- Other related assets.

Use the current plain icon sizes recommended by Microsoft and remove the
1024x1024 and 512x512 px plain Stratoshark application icons. We don't
use them anywhere and they take up space.

Switch Stratoshark.icns to an iconset and add the individual files to a
Qt resource (ssicon_mac.qrc). This ensures that icons look the same in
the Finder and for the running application.

Add StratosharkCap.icns and use them in ssicon_mac.qrc.

Switch Stratoshark's capture start and restart toolbar icons back to the
"fin" variants.