GTP,DIAMETER: IE/AVP="Selection mode" changes
1. GTP: Removed duplicates of value_string `gtp_sel_mode_vals`

3GPP TS 29.060 (7.7.12, 7.7.113) and 3GPP TS 29.274 (8.58) have the
same values.

2. DIAMETER: Reused value_string `gtp_sel_mode_vals` of GTP
In accordance with 3GPP TS 29.061 (page 97)
Selection mode value: Text type.
The format of this sub-attribute shall be a character that represents a single digit,
mapping from the binary value of the selection mode in the
Create PDP Context message (3GPP TS 29.060 [24]) for the GGSN, and the Create Session
Request message (3GPP TS 29.274 [81]) for the P-GW.