cmake: Use in tree libpcap headers for Windows cross-compiling, MSYS2
libpcap is BSD licensed, but Npcap has a non-free license (which
does give Wireshark specific permission to redistribute it.)

Due to the licensing requirements, third party packagers for Windows
will either package a compiled version of libpcap that does not
capture packets on Windows (vcpkg, Conan, MSYS2) or continue to
package the last release of WinPcap (Fedora Linux MinGW.)

On Windows, we compile using the headers from those packages, making
sure not to link the non-functional DLLs, distribute Npcap separately
as part of the installer, and load its wpcap.dll through run-time
dynamic linking, checking for each needed function.

Just add an internal copy of the latest release of the libpcap headers
(1.10.5), and use it unconditionally when Windows cross-compiling or
on MSYS2, or otherwise if libpcap headers aren't found. (On non Windows
we will still require finding the library.) That's all we're actually
doing on Windows and MSYS2 builds as it is; it also makes the Fedora
MinGW builds use the latest API instead of using Npcap with the
libpcap-1.0 / WinPcap 4.1.3 level API.

Don't bother installing the libpcap or WinPcap packages from vcpkg,
MSYS2, and Fedora MinGW, because we don't need them.

Remove the advertising clause from the files with the original
University of California, Berkeley clause, per
  https://ipira.berkeley.edu/sites/default/files/amendment_of_4-clause_bsd_software_license.pdf
originally at
  ftp://ftp.cs.berkeley.edu/pub/4bsd/README.Impt.License.Change

Other files (including pcap.h) have the provision but mentioning the
Computer Systems Engineering Group at LBL; that might still apply
(hopefully it does not.) If it *is* a problem, then it's always been
a problem, because we've always distributed packages including these
headers.

Ping #20261