file: We know compression type at cf_open
Since we know the file type after wtap_open_offline is called, we
must also know the compression type. Otherwise the magic numbers
or heuristics wouldn't work to identify the file type. (We don't
use the extension in either case.) We definitely don't know any
*more* at the top of cf_read than we do in cf_open after calling
wtap_open_offline. However, when reading files from a live capture,
we call cf_open but not cf_read.

So we can retrieve and set it in cf_open and rescan_file, not just
in cf_read, which is necessary in order to record the compression
type of a file if we ever support dumpcap directly writing compressed
files on the fly. (#9311)