DNP3: Reassemble when transport sequence number is reset on FIR bit
IEEE Std 1815-2012 (DNP3) 8.2.1.4 Rules

Rule 4: A transport segment with the FIR bit set may have any sequence
number from 0 to 63 without regard to prior history.

Rule 5 2) A received transport segment having the FIR bit set shall
cause the entire, in-progress transport segment-series to be
discarded...

The full Rule 5 explains that DNP3 is not supposed to handle reordered
transport segments at all. In practice, we can reassemble reordered
segments, but for implementations that take advantage of Rule 4 (likely
resetting the sequence number to 0 for each segment with the FIR bit
set, treating the sequence number as a fragment number), we cannot
reassemble segments that are reordered across a FIR segment.

The alternative implementation, where the sequence number is a rolling.
6 bit counter that increments for each transport segment regardless
of whether it is part of a series, is more robust to re-ordering. We
could have a preference to be able to handle more out-of-order frames
if it is known that the sequence number is such.

Fix #20336