writecap: Remove modelines, normalize spacing
writecap has an .editorconfig file that was copied from another
directory and doesn't refer to its files, so the default EditorConfig
settings are used, which are spaces with tab width 4.
pcapio.c (mostly) has spaces with tab width 8, which the modelines
describe but on editors supporting EditorConfig (our recommended
method) it doesn't work.

Remove the modelines and make the style consistent with our defaults
(spaces with tab width 4 for .[ch], tabs for CMakeList.txt).

Ping #17253