c-ares: work around a problem with static builds.
In an attempt to deal with the import library stuff on Windows,
c-ares set up their CMake build to append "_static" to the
library name when building a static library.

Unfortunately, they did that on all platforms, rather than just
n Windows, so some UN*Xes may have the static libcares named
libcares_static.a rather than libcares.a.

They subsequently fixed that, but on Ubuntu 24.04, as it
currently exists, that fix isn't present, but the Windows-induced
pain is present.

Work around that by looking for cares_static after we look for
cares and libcares-2; that way, we don't get the static library
if we don't require it.

Fixes #20343 (although another static build problem on Ubuntu 24.04
with libpcap popped up, probably due either to the .pc files not
properly listing all dependencies for static linking or due to
pcap-config not doing so).