file_wrappers: Distinguish LZ4 fast seeks to frames vs blocks
Even with independent blocks we have to reset the compression
context when jumping to the start of the same frame. Distinguish
between fast seek points that are the start of LZ4 frames and
those that are at the start of blocks, similar to what is done
with zlib. This is necessary to make "View as File Format" work.