file_wrappers: Make lz4_fill_out_buffer work like zlib_fill_out_buffer
Like zlib_fill_out_buffer, make lz4_fill_out_buffer read up to the
end of a frame or error instead of one block at a time. This makes
for simpler, more parallel code.

This has the effect of causing the file_eof state to be set with a
deferred error when the frame footer is missing (like zlib). That
makes it easier to later add support for reading compressed captures
on the fly if both compression types are read similarly.

Note that we intentionally don't add a fast seek point after the
EndMark, which has no data, and that this prevents calculating the
frame Content Checksum after doing a fast seek, which is good because
unlike zlib the LZ4 Frame API doesn't have a method to update the
running checksum value.