file: Clear EOF in cf_finish_tail
Clear EOF before reading the rest of a live capture for the same reasons
done in cf_continue_tail. This is necessary for any capture type that
might have some extra data added when closed, in particular for compressed
live captures that have a stream or frame footer.

Related to #9311