Qt: Fix extcap interface toolbar Qt6 regression
QComboBox Data is stored as a QVariant. In Qt5, QVariants are converted
for equality comparison if possible, but in Qt6 conversion only happens
in a more limited number of situations. In Qt5, QVariants from a
QByteArray and a QString can compare equal if the QByteArray would
convert to the string, but they never compare equal in Qt6.

The payload passed to the Interface Toolbar is a QByteArray (UTF-8).
Don't prematurely convert it to a QString (UTF-16) because passing it to a
function that converts it to a QVariant, because then it won't compare
equal to QVariants from the same QByteArray not converted.

Also note that Qt5 and Qt6 convert QByteArrays with internal '\0'
differently.

Fix #20354