file_wrappers: Initialize an unread value, check its return
We only call lz4_fill_out_buffer when we're already in a LZ4 frame,
which means that LZ4F_getFrameInfo doesn't read the input, doesn't
read the input size bytes, and always sets it to zero.

Initialize it to make Coverity happy (CID 1641422), and assert that
the return value is zero. It will be if the decompression context
is in the proper stage. Otherwise there's a coding error somewhere.